# docs/apidocs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Cars Management API",
    "description": "## Introduction\n\nThis project is a implementation of a Car Management API in [Node.js](https://nodejs.org/en/). It uses [Express](https://expressjs.com/) as web framework and [PostgreSQL](https://www.postgresql.org/) as database.\n\nThis application allows new users to register and authenticate using the JSON web token. Registered member can view the list of available cars . Admin has additional privileges such as adding new cars to databases or other actions to manage cars & users.",
    "contact": {},
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "variables": {}
    }
  ],
  "paths": {
...
    "/cars/api/v1": {
      "get": {
        "tags": ["Cars"],
        "summary": "Get Cars List",
        "description": "This endpoint is use for displaying a list of available cars, resource can be accessed every user",
        "operationId": "GetCarsList",
        "parameters": [],
        "responses": {
          "default": {
            "description": "",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/m2009"
                    },
                    {
                      "example": {
                        "status": "success",
                        "message": "Successfully got cars data !",
                        "data": [
                          {
                            "id": "43cfb71b-daaf-4680-ba09-ed7a0dd5bb26",
                            "car_name": "Lamborgini Aventador Updated",
                            "car_size": "Large",
                            "car_type": "Sport Car",
                            "car_photo": "/static/1666277306258-901950.jpg",
                            "car_price": 123123112,
                            "car_availability": true,
                            "createdBy": "admin",
                            "updatedBy": "",
                            "deletedBy": "",
                            "createdAt": "2022-10-20T14:48:26.279Z",
                            "updatedAt": "2022-10-20T14:48:26.279Z"
                          },
                          {
                            "id": "8a7451d6-dca5-4ec1-a793-789037cf84e9",
                            "car_name": "Lamborgini Aventador Updated",
                            "car_size": "Large",
                            "car_type": "Sport Car",
                            "car_photo": "/static/1666280948949-pexels-alexgtacar-1592384.jpg",
                            "car_price": 123123112,
                            "car_availability": true,
                            "createdBy": "supermin",
                            "updatedBy": "",
                            "deletedBy": "",
                            "createdAt": "2022-10-20T15:49:08.962Z",
                            "updatedAt": "2022-10-20T15:49:08.962Z"
                          },
                          {
                            "id": "b0981e84-dbcd-481b-92b4-034f9cb69ea6",
                            "car_name": "Lamborgini Aventador Updated",
                            "car_size": "Large",
                            "car_type": "Sport Car",
                            "car_photo": "/static/1666281165305-pexels-alexgtacar-1592384.jpg",
                            "car_price": 123123112,
                            "car_availability": true,
                            "createdBy": "supermin",
                            "updatedBy": "",
                            "deletedBy": "",
                            "createdAt": "2022-10-20T15:52:45.321Z",
                            "updatedAt": "2022-10-20T15:52:45.321Z"
                          },
                          {
                            "id": "50d4d87d-3a6d-4315-a4fa-08274905c4fa",
                            "car_name": "Lamborgini Aventador Updated",
                            "car_size": "Large",
                            "car_type": "Sport Car",
                            "car_photo": "/static/1666281168136-pexels-alexgtacar-1592384.jpg",
                            "car_price": 123123112,
                            "car_availability": true,
                            "createdBy": "supermin",
                            "updatedBy": "",
                            "deletedBy": "",
                            "createdAt": "2022-10-20T15:52:48.145Z",
                            "updatedAt": "2022-10-20T15:52:48.145Z"
                          }
                        ]
                      }
                    }
                  ]
                },
                "example": {
                  "status": "success",
                  "message": "Successfully got cars data !",
                  "data": [
                    {
                      "id": "43cfb71b-daaf-4680-ba09-ed7a0dd5bb26",
                      "car_name": "Lamborgini Aventador Updated",
                      "car_size": "Large",
                      "car_type": "Sport Car",
                      "car_photo": "/static/1666277306258-901950.jpg",
                      "car_price": 123123112,
                      "car_availability": true,
                      "createdBy": "admin",
                      "updatedBy": "",
                      "deletedBy": "",
                      "createdAt": "2022-10-20T14:48:26.279Z",
                      "updatedAt": "2022-10-20T14:48:26.279Z"
                    },
                    {
                      "id": "8a7451d6-dca5-4ec1-a793-789037cf84e9",
                      "car_name": "Lamborgini Aventador Updated",
                      "car_size": "Large",
                      "car_type": "Sport Car",
                      "car_photo": "/static/1666280948949-pexels-alexgtacar-1592384.jpg",
                      "car_price": 123123112,
                      "car_availability": true,
                      "createdBy": "supermin",
                      "updatedBy": "",
                      "deletedBy": "",
                      "createdAt": "2022-10-20T15:49:08.962Z",
                      "updatedAt": "2022-10-20T15:49:08.962Z"
                    },
                    {
                      "id": "b0981e84-dbcd-481b-92b4-034f9cb69ea6",
                      "car_name": "Lamborgini Aventador Updated",
                      "car_size": "Large",
                      "car_type": "Sport Car",
                      "car_photo": "/static/1666281165305-pexels-alexgtacar-1592384.jpg",
                      "car_price": 123123112,
                      "car_availability": true,
                      "createdBy": "supermin",
                      "updatedBy": "",
                      "deletedBy": "",
                      "createdAt": "2022-10-20T15:52:45.321Z",
                      "updatedAt": "2022-10-20T15:52:45.321Z"
                    },
                    {
                      "id": "50d4d87d-3a6d-4315-a4fa-08274905c4fa",
                      "car_name": "Lamborgini Aventador Updated",
                      "car_size": "Large",
                      "car_type": "Sport Car",
                      "car_photo": "/static/1666281168136-pexels-alexgtacar-1592384.jpg",
                      "car_price": 123123112,
                      "car_availability": true,
                      "createdBy": "supermin",
                      "updatedBy": "",
                      "deletedBy": "",
                      "createdAt": "2022-10-20T15:52:48.145Z",
                      "updatedAt": "2022-10-20T15:52:48.145Z"
                    }
                  ]
                }
              }
            }
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "tags": ["Cars"],
        "summary": "Post Cars",
        "description": "This endpoint is use to create a car, resource can only be accessed by admin & super admin.",
        "operationId": "PostCars",
        "parameters": [],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "car_name",
                  "car_size",
                  "car_type",
                  "car_photo",
                  "car_price"
                ],
                "type": "object",
                "properties": {
                  "car_name": {
                    "type": "string",
                    "example": "New Car"
                  },
                  "car_size": {
                    "type": "string",
                    "example": "Large"
                  },
                  "car_type": {
                    "type": "string",
                    "example": "Sport Car"
                  },
                  "car_photo": {
                    "type": "string",
                    "format": "binary"
                  },
                  "car_price": {
                    "type": "integer",
                    "format": "int32",
                    "example": 123123112
                  }
                }
              },
              "encoding": {
                "car_name": {
                  "style": "form"
                },
                "car_size": {
                  "style": "form"
                },
                "car_type": {
                  "style": "form"
                },
                "car_photo": {
                  "style": "form"
                },
                "car_price": {
                  "style": "form"
                }
              }
            }
          },
          "required": false
        },
        "responses": {
          "default": {
            "description": "",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/m20010"
                    },
                    {
                      "example": {
                        "status": "success",
                        "message": "Successfully post a new car with id: 65b73c4e-4de8-4af0-8855-2de9bd3c1724 ",
                        "data": {
                          "id": "65b73c4e-4de8-4af0-8855-2de9bd3c1724",
                          "car_availability": true,
                          "car_name": "New Car",
                          "car_size": "Large",
                          "car_type": "Sport Car",
                          "car_price": 123123112,
                          "car_photo": "/static/1666337181612-pexels-pavlo-luchkovski-337909.jpg",
                          "createdBy": "supermin",
                          "updatedBy": "",
                          "deletedBy": "",
                          "updatedAt": "2022-10-21T07:26:21.639Z",
                          "createdAt": "2022-10-21T07:26:21.639Z"
                        }
                      }
                    }
                  ]
                },
                "example": {
                  "status": "success",
                  "message": "Successfully post a new car with id: 65b73c4e-4de8-4af0-8855-2de9bd3c1724 ",
                  "data": {
                    "id": "65b73c4e-4de8-4af0-8855-2de9bd3c1724",
                    "car_availability": true,
                    "car_name": "New Car",
                    "car_size": "Large",
                    "car_type": "Sport Car",
                    "car_price": 123123112,
                    "car_photo": "/static/1666337181612-pexels-pavlo-luchkovski-337909.jpg",
                    "createdBy": "supermin",
                    "updatedBy": "",
                    "deletedBy": "",
                    "updatedAt": "2022-10-21T07:26:21.639Z",
                    "createdAt": "2022-10-21T07:26:21.639Z"
                  }
                }
              }
            }
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/cars/api/v1/43cfb71b-daaf-4680-ba09-ed7a0dd5bb26": {
      "get": {
        "tags": ["Cars"],
        "summary": "Get Car By Id",
        "description": "This endpoint is use for display a car by id, resource can only be accessed by every user.",
        "operationId": "GetCarById",
        "parameters": [],
        "responses": {
          "default": {
            "description": "",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/m4041"
                    },
                    {
                      "example": {
                        "status": "error",
                        "message": "Error : car not found"
                      }
                    }
                  ]
                },
                "example": {
                  "status": "error",
                  "message": "Error : car not found"
                }
              }
            }
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/cars/api/v1/65b73c4e-4de8-4af0-8855-2de9bd3c1724": {
      "put": {
        "tags": ["Cars"],
        "summary": "Update Cars",
        "description": "This endpoint is use for update a car by id, resource can only be accessed by admin & super admin.",
        "operationId": "UpdateCars",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "encoding": {},
              "schema": {
                "required": [
                  "car_name",
                  "car_size",
                  "car_type",
                  "car_photo",
                  "car_price"
                ],
                "type": "object",
                "properties": {
                  "car_name": {
                    "type": "string",
                    "example": "Car Updated"
                  },
                  "car_size": {
                    "type": "string",
                    "example": "Large"
                  },
                  "car_type": {
                    "type": "string",
                    "example": "Sport Car"
                  },
                  "car_photo": {
                    "type": "string"
                  },
                  "car_price": {
                    "type": "integer",
                    "format": "int32",
                    "example": 123123112
                  }
                }
              }
            }
          },
          "required": false
        },
        "responses": {
          "default": {
            "description": "",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/m20010"
                    },
                    {
                      "example": {
                        "status": "success",
                        "message": "Successfully post a new car with id: 65b73c4e-4de8-4af0-8855-2de9bd3c1724 ",
                        "data": {
                          "id": "65b73c4e-4de8-4af0-8855-2de9bd3c1724",
                          "car_name": "Car Updated",
                          "car_size": "Large",
                          "car_type": "Sport Car",
                          "car_photo": "/static/1666337362918-pexels-alexgtacar-1592384.jpg",
                          "car_price": 123123112,
                          "car_availability": true,
                          "createdBy": "supermin",
                          "updatedBy": "supermin",
                          "deletedBy": "",
                          "createdAt": "2022-10-21T07:26:21.639Z",
                          "updatedAt": "2022-10-21T07:29:22.997Z"
                        }
                      }
                    }
                  ]
                },
                "example": {
                  "status": "success",
                  "message": "Successfully post a new car with id: 65b73c4e-4de8-4af0-8855-2de9bd3c1724 ",
                  "data": {
                    "id": "65b73c4e-4de8-4af0-8855-2de9bd3c1724",
                    "car_name": "Car Updated",
                    "car_size": "Large",
                    "car_type": "Sport Car",
                    "car_photo": "/static/1666337362918-pexels-alexgtacar-1592384.jpg",
                    "car_price": 123123112,
                    "car_availability": true,
                    "createdBy": "supermin",
                    "updatedBy": "supermin",
                    "deletedBy": "",
                    "createdAt": "2022-10-21T07:26:21.639Z",
                    "updatedAt": "2022-10-21T07:29:22.997Z"
                  }
                }
              }
            }
          }
        },
        "deprecated": false
      },
      "delete": {
        "tags": ["Cars"],
        "summary": "Remove Car",
        "description": "This endpoint is use for remove car by id, resource can only be accessed by admin & super admin.",
        "operationId": "RemoveCar",
        "parameters": [],
        "responses": {
          "default": {
            "description": "",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/m20010"
                    },
                    {
                      "example": {
                        "status": "success",
                        "message": "Successfully deleted car with id : 65b73c4e-4de8-4af0-8855-2de9bd3c1724",
                        "data": {
                          "id": "65b73c4e-4de8-4af0-8855-2de9bd3c1724",
                          "car_name": "Car Updated",
                          "car_size": "Large",
                          "car_type": "Sport Car",
                          "car_photo": "/static/1666337362918-pexels-alexgtacar-1592384.jpg",
                          "car_price": 123123112,
                          "car_availability": false,
                          "createdBy": "supermin",
                          "updatedBy": "supermin",
                          "deletedBy": "supermin",
                          "createdAt": "2022-10-21T07:26:21.639Z",
                          "updatedAt": "2022-10-21T07:32:50.143Z"
                        }
                      }
                    }
                  ]
                },
                "example": {
                  "status": "success",
                  "message": "Successfully deleted car with id : 65b73c4e-4de8-4af0-8855-2de9bd3c1724",
                  "data": {
                    "id": "65b73c4e-4de8-4af0-8855-2de9bd3c1724",
                    "car_name": "Car Updated",
                    "car_size": "Large",
                    "car_type": "Sport Car",
                    "car_photo": "/static/1666337362918-pexels-alexgtacar-1592384.jpg",
                    "car_price": 123123112,
                    "car_availability": false,
                    "createdBy": "supermin",
                    "updatedBy": "supermin",
                    "deletedBy": "supermin",
                    "createdAt": "2022-10-21T07:26:21.639Z",
                    "updatedAt": "2022-10-21T07:32:50.143Z"
                  }
                }
              }
            }
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/cars/carsDump": {
      "get": {
        "tags": ["Cars"],
        "summary": "Get Cars Dump",
        "description": "This endpoint is use display unavailable cars, resource can only be accessed by admin & super admin.",
        "operationId": "GetCarsDump",
        "parameters": [],
        "responses": {
          "default": {
            "description": "",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/m2009"
                    },
                    {
                      "example": {
                        "status": "success",
                        "message": "Successfully got cars data !",
                        "data": [
                          {
                            "id": "3f3961b3-e3a5-4dac-af25-d6ce4df36696",
                            "car_name": "UPDATED CARRRRR",
                            "car_size": "Large",
                            "car_type": "Sport Car",
                            "car_photo": "/static/1666281412547-pexels-albin-berlin-296735.jpg",
                            "car_price": 123123112,
                            "car_availability": false,
                            "createdBy": "admin",
                            "updatedBy": "supermin",
                            "deletedBy": "supermin",
                            "createdAt": "2022-10-20T14:46:47.906Z",
                            "updatedAt": "2022-10-20T16:07:34.136Z"
                          },
                          {
                            "id": "65b73c4e-4de8-4af0-8855-2de9bd3c1724",
                            "car_name": "Car Updated",
                            "car_size": "Large",
                            "car_type": "Sport Car",
                            "car_photo": "/static/1666337362918-pexels-alexgtacar-1592384.jpg",
                            "car_price": 123123112,
                            "car_availability": false,
                            "createdBy": "supermin",
                            "updatedBy": "supermin",
                            "deletedBy": "supermin",
                            "createdAt": "2022-10-21T07:26:21.639Z",
                            "updatedAt": "2022-10-21T07:32:50.143Z"
                          }
                        ]
                      }
                    }
                  ]
                },
                "example": {
                  "status": "success",
                  "message": "Successfully got cars data !",
                  "data": [
                    {
                      "id": "3f3961b3-e3a5-4dac-af25-d6ce4df36696",
                      "car_name": "UPDATED CARRRRR",
                      "car_size": "Large",
                      "car_type": "Sport Car",
                      "car_photo": "/static/1666281412547-pexels-albin-berlin-296735.jpg",
                      "car_price": 123123112,
                      "car_availability": false,
                      "createdBy": "admin",
                      "updatedBy": "supermin",
                      "deletedBy": "supermin",
                      "createdAt": "2022-10-20T14:46:47.906Z",
                      "updatedAt": "2022-10-20T16:07:34.136Z"
                    },
                    {
                      "id": "65b73c4e-4de8-4af0-8855-2de9bd3c1724",
                      "car_name": "Car Updated",
                      "car_size": "Large",
                      "car_type": "Sport Car",
                      "car_photo": "/static/1666337362918-pexels-alexgtacar-1592384.jpg",
                      "car_price": 123123112,
                      "car_availability": false,
                      "createdBy": "supermin",
                      "updatedBy": "supermin",
                      "deletedBy": "supermin",
                      "createdAt": "2022-10-21T07:26:21.639Z",
                      "updatedAt": "2022-10-21T07:32:50.143Z"
                    }
                  ]
                }
              }
            }
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "m4041": {
        "title": "m4041",
        "required": ["status", "message"],
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "example": {
          "status": "error",
          "message": "Username or password is incorrect"
        }
      },
...
      "m2009": {
        "title": "m2009",
        "required": ["status", "message", "data"],
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Data8"
            },
            "description": ""
          }
        },
        "example": {
          "status": "success",
          "message": "Successfully got cars data !",
          "data": [
            {
              "id": "43cfb71b-daaf-4680-ba09-ed7a0dd5bb26",
              "car_name": "Lamborgini Aventador Updated",
              "car_size": "Large",
              "car_type": "Sport Car",
              "car_photo": "/static/1666277306258-901950.jpg",
              "car_price": 123123112,
              "car_availability": true,
              "createdBy": "admin",
              "updatedBy": "",
              "deletedBy": "",
              "createdAt": "2022-10-20T14:48:26.279Z",
              "updatedAt": "2022-10-20T14:48:26.279Z"
            },
            {
              "id": "8a7451d6-dca5-4ec1-a793-789037cf84e9",
              "car_name": "Lamborgini Aventador Updated",
              "car_size": "Large",
              "car_type": "Sport Car",
              "car_photo": "/static/1666280948949-pexels-alexgtacar-1592384.jpg",
              "car_price": 123123112,
              "car_availability": true,
              "createdBy": "supermin",
              "updatedBy": "",
              "deletedBy": "",
              "createdAt": "2022-10-20T15:49:08.962Z",
              "updatedAt": "2022-10-20T15:49:08.962Z"
            },
            {
              "id": "b0981e84-dbcd-481b-92b4-034f9cb69ea6",
              "car_name": "Lamborgini Aventador Updated",
              "car_size": "Large",
              "car_type": "Sport Car",
              "car_photo": "/static/1666281165305-pexels-alexgtacar-1592384.jpg",
              "car_price": 123123112,
              "car_availability": true,
              "createdBy": "supermin",
              "updatedBy": "",
              "deletedBy": "",
              "createdAt": "2022-10-20T15:52:45.321Z",
              "updatedAt": "2022-10-20T15:52:45.321Z"
            },
            {
              "id": "50d4d87d-3a6d-4315-a4fa-08274905c4fa",
              "car_name": "Lamborgini Aventador Updated",
              "car_size": "Large",
              "car_type": "Sport Car",
              "car_photo": "/static/1666281168136-pexels-alexgtacar-1592384.jpg",
              "car_price": 123123112,
              "car_availability": true,
              "createdBy": "supermin",
              "updatedBy": "",
              "deletedBy": "",
              "createdAt": "2022-10-20T15:52:48.145Z",
              "updatedAt": "2022-10-20T15:52:48.145Z"
            }
          ]
        }
      },
      "Data8": {
        "title": "Data8",
        "required": [
          "id",
          "car_name",
          "car_size",
          "car_type",
          "car_photo",
          "car_price",
          "car_availability",
          "createdBy",
          "updatedBy",
          "deletedBy",
          "createdAt",
          "updatedAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "car_name": {
            "type": "string"
          },
          "car_size": {
            "type": "string"
          },
          "car_type": {
            "type": "string"
          },
          "car_photo": {
            "type": "string"
          },
          "car_price": {
            "type": "integer",
            "format": "int32"
          },
          "car_availability": {
            "type": "boolean"
          },
          "createdBy": {
            "type": "string"
          },
          "updatedBy": {
            "type": "string"
          },
          "deletedBy": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "example": {
          "id": "43cfb71b-daaf-4680-ba09-ed7a0dd5bb26",
          "car_name": "Lamborgini Aventador Updated",
          "car_size": "Large",
          "car_type": "Sport Car",
          "car_photo": "/static/1666277306258-901950.jpg",
          "car_price": 123123112,
          "car_availability": true,
          "createdBy": "admin",
          "updatedBy": "",
          "deletedBy": "",
          "createdAt": "2022-10-20T14:48:26.279Z",
          "updatedAt": "2022-10-20T14:48:26.279Z"
        }
      },
      "m20010": {
        "title": "m20010",
        "required": ["status", "message", "data"],
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/Data8"
          }
        },
        "example": {
          "status": "success",
          "message": "Successfully got car data !",
          "data": {
            "id": "43cfb71b-daaf-4680-ba09-ed7a0dd5bb26",
            "car_name": "Lamborgini Aventador Updated",
            "car_size": "Large",
            "car_type": "Sport Car",
            "car_photo": "/static/1666277306258-901950.jpg",
            "car_price": 123123112,
            "car_availability": true,
            "createdBy": "admin",
            "updatedBy": "",
            "deletedBy": "",
            "createdAt": "2022-10-20T14:48:26.279Z",
            "updatedAt": "2022-10-20T14:48:26.279Z"
          }
        }
      }
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [],
...
}